commit20220212-02
https://app.asana.com/0/1201715872024270/1201809862024115
https://app.asana.com/0/1201715872024270/1201738155714321

+++ CMS_3D_Core/CMS_3D_Core/Models/EDM/t_light.cs
@@ -12,7 +12,7 @@ namespace CMS_3D_Core.Models.EDM
         public string light_type { get; set; }
         public string title { get; set; }
         public string short_description { get; set; }
-        public int? color { get; set; }
+        public long? color { get; set; }
         public double? intensity { get; set; }
         public double? px { get; set; }
         public double? py { get; set; }
@@ -24,8 +24,8 @@ namespace CMS_3D_Core.Models.EDM
         public double? tx { get; set; }
         public double? ty { get; set; }
         public double? tz { get; set; }
-        public int? skycolor { get; set; }
-        public int? groundcolor { get; set; }
+        public long? skycolor { get; set; }
+        public long? groundcolor { get; set; }
         public bool is_lensflare { get; set; }
         public double? lfsize { get; set; }
         public byte[] file_data { get; set; }
